trading-notes-platform/trading-notes-api/requirements.txt
Flask==2.3.3
google-cloud-firestore==2.11.0
firebase-admin==6.2.0
gunicorn==21.2.0
